Serve web game data from Spring with cache
